implement validator and parser

+++ src/main/java/org/ibanfirst/trade/model/Trade.java
@@ -15,9 +15,9 @@ public class Trade {
     private LocalDate tradeDate;
     private String symbol;
     private String currency;
-    private Double amount;
-    private Double amountCounterValue;
-    private Double rate;
+    private Float amount;
+    private Float amountCounterValue;
+    private Float rate;
     private LocalDate valueDate;
